requirements.txt
# Core dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
